+++ src/test/java/kr/hhplus/be/server/application/product/PopularProductCacheTest.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.server.application.product;
 
-import kr.hhplus.be.server.config.TestRedisCacheConfig;
+import kr.hhplus.be.server.config.redis.TestRedisCacheConfig;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
